pyproject.toml
[tool.poetry]
name = "airflow_docker_pineline"
version = "0.1.0"
description = ""
authors = ["bijeck"]

[tool.poetry.dependencies]
python = "^3.9"
mysql-connector-python = "^8.0.29"
pylint = "^2.14.3"
sorcery = "^0.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
